+++ CHANGELOG.md
@@ -20,6 +20,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 - If any Post or Page content contains images with class of 'lazyload' you'll need to use the `#editor_body` method when instaniating Ckeditor otherwise the lazyload images will not appear within the editor
 - All Posts must be associated to an Integral::Category through the `category` association
+- Added User statuses - pending, active & blocked
+- User managers can now block other users from logging in and accessing any backend functionality
 
 ### Added